+++ src/main/java/com/sdjr2/rest_contact_meanssb/repositories/AddressJpaRepository.java
@@ -22,7 +22,7 @@ import java.util.Optional;
  * @since 23/06/10
  */
 @Repository
-public interface AddressJpaRepository extends JpaRepository<AddressEntity, Integer>,
+public interface AddressJpaRepository extends JpaRepository<AddressEntity, Long>,
 		JpaSpecificationExecutor<AddressEntity> {
 
 	@Query("SELECT DISTINCT a.town FROM AddressEntity a")
